requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==70.1.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
